+++ huskarl/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changes
+
+ - Breaking: Update HttpAuthorizer to accept a GrantTokenSource which is capable
+   of regenerating grant parameters (e.g. useful for JWT bearer grant). Also this
+   unifies with the idea of priming the token source.
+
 ## [0.8.0] - 2026-06-15
 
 ### Changes